+++ packages/dev/src/App.vue
@@ -3,9 +3,9 @@
 </template>
 
 <script setup lang="ts">
-import "@simsustech/quasar-components/css";
-import { provide, ref } from "vue";
-import { useQuasar } from "quasar";
+import '@simsustech/quasar-components/css'
+import { provide, ref } from 'vue'
+import { useQuasar } from 'quasar'
 </script>
 
 <style lang="sass">