+++ packages/dev/src/pages/Index.vue
@@ -5,5 +5,5 @@
 </template>
 
 <script setup lang="ts">
-import HelloWorld from "src/components/HelloWorld.vue";
+import HelloWorld from 'src/components/HelloWorld.vue'
 </script>