+++ packages/dev/src/router/routes.ts
@@ -1,9 +1,9 @@
-import { h } from "vue";
-import { RouteRecordRaw } from "vue-router";
+import { h } from 'vue'
+import { RouteRecordRaw } from 'vue-router'
 
-const authenticationPages = import.meta.glob("../pages/authentication/*.vue");
-const generalPages = import.meta.glob("../pages/general/*.vue");
-const formPages = import.meta.glob("../pages/form/*.vue");
+const authenticationPages = import.meta.glob('../pages/authentication/*.vue')
+const generalPages = import.meta.glob('../pages/general/*.vue')
+const formPages = import.meta.glob('../pages/form/*.vue')
 // export const authenticationRoutes = Object.entries(authenticationPages).reduce(
 //   (acc, [key, val]) => {
 //     const name = key.split("/").at(-1)?.split(".").at(0);
@@ -21,21 +21,21 @@ const formPages = import.meta.glob("../pages/form/*.vue");
 export const getRoutes = (pages: Record<string, () => any>) =>
   Object.entries(pages).reduce(
     (acc, [key, val]) => {
-      const name = key.split("/").at(-1)?.split(".").at(0);
+      const name = key.split('/').at(-1)?.split('.').at(0)
       if (name)
         acc.push({
           name,
           path: name.toLowerCase(),
-          component: val,
-        });
-      return acc;
+          component: val
+        })
+      return acc
     },
     [] as { name: string; path: string; component: any }[]
-  );
+  )
 
-export const authenticationRoutes = getRoutes(authenticationPages);
-export const generalRoutes = getRoutes(generalPages);
-export const formRoutes = getRoutes(formPages);
+export const authenticationRoutes = getRoutes(authenticationPages)
+export const generalRoutes = getRoutes(generalPages)
+export const formRoutes = getRoutes(formPages)
 
 // const testPages = import.meta.glob("../pages/test/*.vue");
 
@@ -52,42 +52,42 @@ export const formRoutes = getRoutes(formPages);
 
 const routes: RouteRecordRaw[] = [
   {
-    path: "/",
-    component: () => import("src/layouts/MainLayout.vue"),
+    path: '/',
+    component: () => import('src/layouts/MainLayout.vue'),
     children: [
-      { path: "", component: () => import("src/pages/Index.vue") },
+      { path: '', component: () => import('src/pages/Index.vue') },
       {
-        path: "general",
-        component: () => import("../pages/General.vue"),
-        children: generalRoutes,
+        path: 'general',
+        component: () => import('../pages/General.vue'),
+        children: generalRoutes
       },
       {
-        path: "authentication",
-        component: () => import("../pages/Authentication.vue"),
-        children: authenticationRoutes,
+        path: 'authentication',
+        component: () => import('../pages/Authentication.vue'),
+        children: authenticationRoutes
       },
       {
-        path: "form",
-        component: () => import("../pages/Form.vue"),
-        children: formRoutes,
+        path: 'form',
+        component: () => import('../pages/Form.vue'),
+        children: formRoutes
       },
       {
-        path: "flags",
-        component: () => import("../pages/Flags.vue"),
+        path: 'flags',
+        component: () => import('../pages/Flags.vue')
       },
       {
-        path: "icons",
-        component: () => import("../pages/Icons.vue"),
-      },
+        path: 'icons',
+        component: () => import('../pages/Icons.vue')
+      }
       // ...testRoutes,
-    ],
+    ]
   },
   // Always leave this as last one,
   // but you can also remove it
   {
-    path: "/:catchAll(.*)*",
-    component: () => import("src/pages/Error404.vue"),
-  },
-];
+    path: '/:catchAll(.*)*',
+    component: () => import('src/pages/Error404.vue')
+  }
+]
 
-export default routes;
+export default routes